+++ src/lib/common/SimpleAllocator.cpp
@@ -326,7 +326,7 @@ void SimpleAllocator::touchMemory(void* ptr, size_t size)
 {
 	if (ptr == NULL || ptr == MAP_FAILED)
 		return;
-	for (int i = 0 ; i < size ; i += MATT_PAGE_SIZE)
+	for (size_t i = 0 ; i < size ; i += MATT_PAGE_SIZE)
 		*(char*)ptr = '\0';
 }
